break down app to components
