+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { Container } from "@mui/material";
-import TodoCard from "./components/TodoCard";
+import TodoList from "./components/TodoList";
 import { useEffect, useState } from "react";
 import axios from "axios";
 import Header from "./components/Header";
@@ -11,12 +11,12 @@ function App() {
     axios
       .get("http://localhost:3000/todos")
       .then((response) => setTodos(response.data));
-  }, []);
+  }, [todos]);
 
   return (
     <Container>
       <Header></Header>
-      <TodoCard todos={todos}></TodoCard>
+      <TodoList todoList={todos}></TodoList>
     </Container>
   );
 }
